chore: formatting

+++ src/main/java/rendering/Renderer.java
@@ -12,8 +12,11 @@ import org.joml.Matrix4f;
  * Main renderer responsible for drawing each frame.
  * Handles all OpenGL draw calls including such as terrain, path and camera.
  * 
- * <p>Uses a single {@link Shader} program for both terrain and path rendering. Path is rendered
- * in a separate color defined by the {@code pathColor} uniform.</p>
+ * <p>
+ * Uses a single {@link Shader} program for both terrain and path rendering.
+ * Path is rendered
+ * in a separate color defined by the {@code pathColor} uniform.
+ * </p>
  */
 public class Renderer {
     /** Shader program for rendering terrain and path */
@@ -31,24 +34,26 @@ public class Renderer {
     /**
      * Constructs a new {@code Renderer}.
      * 
-     * @param camera Camera providing view and projection matrices
+     * @param camera       Camera providing view and projection matrices
      * @param chunkManager Chunk manager providing visible terrain chunks
-     * @param aStar A* instance for current path
+     * @param aStar        A* instance for current path
      */
-    public Renderer(Camera camera, ChunkManager chunkManager, AStar aStar) {}
+    public Renderer(Camera camera, ChunkManager chunkManager, AStar aStar) {
+    }
 
     /**
      * Initializes the shader program. Must be called after OpenGL context
      * is created and before first call to {@link #render(float)}!
      */
-    public void init() {}
+    public void init() {
+    }
 
     /**
      * Renders a single frame, called once per iteration of the application loop.
      * 
      * @param deltaTime Time in seconds since the last render
      */
-    public void render(float deltaTime) {
+    public void render(double deltaTime) {
         // Clear framebuffer
         glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
         renderTerrain();
@@ -58,12 +63,14 @@ public class Renderer {
     /**
      * Renders all visible terrain chunks provided by {@link ChunkManager}.
      */
-    private void renderTerrain() {}
+    private void renderTerrain() {
+    }
 
     /**
      * Renders the current path provided by {@link AStar#getPath()}.
      */
-    private void renderPath() {}
+    private void renderPath() {
+    }
 
     /**
      * Builds the model-view-projection matrix for a given model transformation.
@@ -71,11 +78,13 @@ public class Renderer {
      * @param model Model matrix of the object being rendered
      * @return Combined MVP matrix ({@code projection * view * model})
      */
-    private Matrix4f buildMVP(Matrix4f model) {}
+    private Matrix4f buildMVP(Matrix4f model) {
+    }
 
     /**
      * Releases all OpenGL resources held by the renderer.
      * Called by {@link Application#cleanUp()} on shutdown.
      */
-    public void cleanUp() {}
+    public void cleanUp() {
+    }
 }
